separate ConstructorChecker
