+++ src/main/kotlin/com/vasc/checks/StatementType.kt
@@ -1,0 +1,5 @@
+package com.vasc.checks
+
+enum class StatementType {
+    EXHAUSTIVE_RETURN, NOT_EXHAUSTIVE_RETURN, THIS_CONSTRUCTOR, THIS_CALL, SUPER, OTHER
+}
