+++ src/main/kotlin/com/vasc/checks/constructor/ConstructorChecker.kt
@@ -1,23 +1,20 @@
-package com.vasc.exhaustiveness
+package com.vasc.checks.constructor
 
 import com.vasc.VascTypeResolver
 import com.vasc.antlr.VascParser.*
 import com.vasc.antlr.VascParserBaseVisitor
+import com.vasc.checks.*
 import com.vasc.member.VascConstructor
 import com.vasc.type.VascType
 import org.antlr.v4.runtime.ParserRuleContext
 
-class ExhaustivenessChecker(
+class ConstructorChecker(
     private val typeResolver: VascTypeResolver,
     private val typeTable: MutableMap<ParserRuleContext, VascType>
-) : VascParserBaseVisitor<ExhaustivenessChecker.StatementType>() {
-
-    private var waitReturn = true
+) : VascParserBaseVisitor<StatementType>() {
     private var currentConstructor: VascConstructor? = null
     private lateinit var currentClass: VascType
     private var constructorCalls: MutableMap<VascConstructor, VascConstructor> = mutableMapOf()
-
-    private val completeIf = 2
 
     override fun visitClassDeclaration(ctx: ClassDeclarationContext): StatementType {
         currentClass = ctx.name.accept(typeResolver)
@@ -27,11 +24,6 @@ class ExhaustivenessChecker(
 
         constructorCalls = mutableMapOf()
         return super.visitClassDeclaration(ctx)
-    }
-
-    override fun visitMethodDeclaration(ctx: MethodDeclarationContext): StatementType {
-        waitReturn = ctx.returnType != null
-        return super.visitMethodDeclaration(ctx)
     }
 
     override fun visitConstructorDeclaration(ctx: ConstructorDeclarationContext): StatementType {
@@ -50,14 +42,11 @@ class ExhaustivenessChecker(
         val statements = ctx.statement().toList()
         val n: Int = statements.size
 
-        val parentIsMethod = ctx.parent is MethodDeclarationContext
+        ctx.parent is MethodDeclarationContext
 
         for (i in 0 until n) {
             statementType = statements[i].accept(this)
             when (statementType) {
-                StatementType.EXHAUSTIVE_RETURN ->
-                    if (i != n - 1) throw UnreachableCodeException("Unreachable code (line ${statements[i + 1].start.line})")
-                    else if (parentIsMethod) waitReturn = false
 
                 StatementType.THIS_CONSTRUCTOR -> {
                     if (currentConstructor == null || i != 0)
@@ -75,8 +64,6 @@ class ExhaustivenessChecker(
             }
         }
 
-        if (parentIsMethod && waitReturn)
-            throw NonExhaustiveReturnException("Nonexhaustive return (line ${ctx.start.line})")
         if (currentConstructor != null && !wasSuperOrThisCall && !currentClass.parentHasDefaultConstructor())
             throw DefaultConstructorNotExistException("Default constructor not exist. Use super call. (line ${ctx.start.line})")
         return statementType
@@ -90,30 +77,13 @@ class ExhaustivenessChecker(
         return StatementType.THIS_CONSTRUCTOR
     }
 
-    override fun visitIfStatement(ctx: IfStatementContext): StatementType {
-        var returnCount = 0
-        ctx.children.forEach {
-            val childResult = it.accept(this)
-            if (childResult == StatementType.EXHAUSTIVE_RETURN) returnCount++
-        }
-
-        return if (returnCount == completeIf) StatementType.EXHAUSTIVE_RETURN
-        else StatementType.NOT_EXHAUSTIVE_RETURN
-    }
-
-
-    override fun visitReturnStatement(ctx: ReturnStatementContext): StatementType {
-        super.visitReturnStatement(ctx)
-        return StatementType.EXHAUSTIVE_RETURN
-    }
-
     override fun visitSuperExpression(ctx: SuperExpressionContext?): StatementType {
         super.visitSuperExpression(ctx)
         return StatementType.SUPER
     }
 
-    override fun defaultResult(): StatementType {
-        return StatementType.OTHER
+    private fun VascType.parentHasDefaultConstructor(): Boolean {
+        return parent == null || parent!!.getDeclaredConstructor(emptyList()) != null || parent!!.declaredConstructors.isEmpty()
     }
 
     private fun getConstructorByArgs(args: ArgumentsContext): VascConstructor {
@@ -130,12 +100,7 @@ class ExhaustivenessChecker(
         }
     }
 
-    private fun VascType.parentHasDefaultConstructor(): Boolean {
-        return parent == null || parent!!.getDeclaredConstructor(emptyList()) != null || parent!!.declaredConstructors.isEmpty()
-    }
-
-    enum class StatementType {
-        EXHAUSTIVE_RETURN, NOT_EXHAUSTIVE_RETURN, THIS_CONSTRUCTOR, THIS_CALL, SUPER, OTHER
+    override fun defaultResult(): StatementType {
+        return StatementType.OTHER
     }
 }
-
